Changes in Reservations/Create view, ReservationViewModel and ReservationsController
Forgot to take out name="guestName"; now its out and it goes to the conroller

Added names of the ReservationViewModel to the Create view.

Made instances of guest and reservation in the controller.  Added the input of the view to those objects. Gave instance of type Reservation back to return View. Could see the date, hours, min, hidePrices, Comments in the Reservations/Index page.

Made some adjustments in the site.js, because were listed in the Error list while running (though it worked with those errors)